behind-ai/behind-flask/requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dlib==19.24.0
executing==1.2.0
ez-setup==0.9
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
imageio==2.25.0
importlib-metadata==6.0.0
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
networkx==3.0
ninja==1.11.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==3.19.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tifffile==2023.2.3
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.12.0
